Create Overview
FUNCTIONS
A function is a group of statements performing a specific task.
When a program gets bigger in size and its complexity grows, it gets difficult for a program to keep track on which piece of code is doing what!

EXAMPLE AND SYNTAX OF A FUNCTION
The syntax of a function looks as follows:
def func1():
    print('hello')
#This function can be called any number of times, anywhere in the program.

FUNCTION CALL
Whenever we want to call a function, we put the name of the function followed by parentheses as follows:
func1() # This is called function call.

FUNCTION DEFINITION
The part containing the exact set of instructions which are executed during the function call.

TYPES OF FUNCTIONS IN PYTHON
There are two types of functions in python:
• Built in functions (Already present in python)
• User defined functions (Defined by the user)

FUNCTIONS WITH ARGUMENTS
A function can accept some value it can work with. We can put these values in the parentheses.

DEFAULT PARAMETER VALUE
We can have a value as default as default argument in a function.
If we specify name = “stranger” in the line containing def, this value is used when no argument is passed.

RECURSION
Recursion is a function which calls itself.
It is used to directly use a mathematical formula as function. 

#NOTE: The programmer needs to be extremely careful while working with recursion to ensure
that the function doesn’t infinitely keep calling itself. Recursion is sometimes the most
direct way to code an algorithm.